attached_proxy_web: consolidate sign-relay into single endpoint
